Refactor: Repository 추상화
- Repository를 추상화하기 위해 공통부분으로 이동하였다.
(Repository의 기능이 대부분 공통으로 사용되므로 추가해야할 메서드가 있는 부분에서만 인터페이스를 상속받는 Repository를 새로 만들면 될듯)
- 이동에 따른 코드 수정

Related to: #37
